requirements.txt
fastapi
uvicorn
firebase-admin
google-genai
python-dotenv
dateparser
requests